requirements.txt
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
distro==1.9.0
flask==3.1.3
flask-cors==6.0.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.13.0
markupsafe==3.0.3
numpy==2.4.3
openai==2.26.0
pandas==3.0.1
pydantic==2.12.5
pydantic-core==2.41.5
pymupdf==1.27.2
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
werkzeug==3.1.6
python-dotenv